tests/parity/python/pyproject.toml
[project]
name = "dia-parity-reference"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
  "pyannote.audio >= 3.1",
  "pyannote.metrics >= 3.2",
]
